nq-downloader/pyproject.toml
[tool.poetry]
name = "nq-downloader"
version = "0.1.0"
description = "Natural Questions dataset downloader for Vertex AI search"
authors = ["Genesis <genesis@local>"]
readme = "README.md"
packages = [{include = "nq_downloader", from = "src"}]

[tool.poetry.scripts]
nq-downloader = "nq_downloader.main:main"

[tool.poetry.dependencies]
python = "^3.13"
# Genesis shared utilities will be available once published to PyPI
# shared-core = "^0.1.0"  # ConfigLoader, get_logger, retry, health checks, etc.
click = "^8.1.7"
pydantic = "^2.5.0"
rich = "^13.7.0"
python-dotenv = "^1.0.0"  # For environment variable loading
google-cloud-storage = "^2.10.0"  # For GCS downloads
google-api-core = "^2.11.0"  # For retry logic

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.1"
ruff = "^0.1.0"
pre-commit = "^3.6.0"
# Note: Genesis CLI is automatically installed during bootstrap

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["nq_downloader"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
pythonpath = ["src"]
addopts = [
    "--cov=nq_downloader",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80"
]
